% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine_fun.R
\name{wine_fun}
\alias{wine_fun}
\title{Plot function}
\usage{
wine_fun(type = "linear", growth = 1)
}
\arguments{
\item{type}{Linear or exponential? Default is type = 'linear' (options are type='linear' or type='exponential')}

\item{growth}{How much does wine effect happiness? Default is growth = 1}
}
\description{
This function generates a plot with linear or exponential growth
}
\examples{
wine_fun(type='exponential', growth='5')
wine_fun()
}
\keyword{wine}
